- Lots of changes to include appointments, products, user companies, etc

+++ src/main/java/com/managedagents/stateless/OrdersBean.java
@@ -25,68 +25,56 @@ import org.primefaces.model.SortOrder;
  * @author james
  */
 @Stateless
-public class OrdersBean
-{
+public class OrdersBean {
+
     public static final String ORDER_ID = "orderId";
-    
+
     @PersistenceContext(unitName = "ManagedAgentsPU")
     private EntityManager em;
-    
-    public List<Orders> findAllOrders()
-    {
+
+    public List<Orders> findAllOrders() {
         TypedQuery<Orders> query = em.createNamedQuery("Orders.findAll", Orders.class);
         return query.getResultList();
     }
 
-    public List<Orders> findUserOrders(Users user)
-    {
+    public List<Orders> findUserOrders(Users user) {
         TypedQuery<Orders> query = em.createNamedQuery("Orders.findByUser", Orders.class);
         query.setParameter("user", user);
         return query.getResultList();
     }
 
-    public Orders findOrderById(Integer orderId)
-    {
+    public Orders findOrderById(Integer orderId) {
         TypedQuery<Orders> query = em.createNamedQuery("Orders.findByOrderId", Orders.class);
         query.setParameter(ORDER_ID, orderId);
-        if (query.getResultList().isEmpty())
-        {
+        if (query.getResultList().isEmpty()) {
             return null;
         }
-        else
-        {
+        else {
             return query.getResultList().get(0);
         }
     }
-    
-    public Long countAllOrders(String sortField, SortOrder sortOrder, Map<String, Object> filters)
-    {
+
+    public Long countAllOrders(String sortField, SortOrder sortOrder, Map<String, Object> filters) {
         CriteriaBuilder cb = em.getCriteriaBuilder();
         CriteriaQuery<Long> cq = cb.createQuery(Long.class);
         Root<Orders> order = cq.from(Orders.class);
-        
-        if (sortField == null)
-        {
+
+        if (sortField == null) {
             sortField = ORDER_ID;
         }
-        
-        if (sortOrder.equals(SortOrder.ASCENDING))
-        {
+
+        if (sortOrder.equals(SortOrder.ASCENDING)) {
             cq.orderBy(cb.asc(order.get(sortField)));
         }
-        else if (sortOrder.equals(SortOrder.DESCENDING))
-        {
+        else if (sortOrder.equals(SortOrder.DESCENDING)) {
             cq.orderBy(cb.desc(order.get(sortField)));
         }
         List<Predicate> predicateList = new ArrayList<>();
-        for (Map.Entry<String, Object> filter : filters.entrySet())
-        {
-            if (filter.getKey().equals("orderStatus") || filter.getKey().equals(ORDER_ID))
-            {
+        for (Map.Entry<String, Object> filter : filters.entrySet()) {
+            if (filter.getKey().equals("orderStatus") || filter.getKey().equals(ORDER_ID)) {
                 predicateList.add(cb.equal(order.get(filter.getKey()), filter.getValue().toString()));
             }
-            else
-            {
+            else {
                 predicateList.add(cb.like(order.get(filter.getKey()), "%" + filter.getValue().toString() + "%"));
             }
         }
@@ -95,38 +83,31 @@ public class OrdersBean
 
         cq.where(predicateArray);
         cq.select(cb.count(order));
-        return em.createQuery(cq).getSingleResult();        
+        return em.createQuery(cq).getSingleResult();
     }
-    
-    public List<Orders> findAllOrders(int first, int pageSize, String sortField, SortOrder sortOrder, Map<String, Object> filters)
-    {
+
+    public List<Orders> findAllOrders(int first, int pageSize, String sortField, SortOrder sortOrder, Map<String, Object> filters) {
         CriteriaBuilder cb = em.getCriteriaBuilder();
         CriteriaQuery<Orders> cq = cb.createQuery(Orders.class);
         Root<Orders> order = cq.from(Orders.class);
         CriteriaQuery<Orders> select = cq.select(order);
-        
-        if (sortField == null)
-        {
+
+        if (sortField == null) {
             sortField = ORDER_ID;
         }
-        
-        if (sortOrder.equals(SortOrder.ASCENDING))
-        {
+
+        if (sortOrder.equals(SortOrder.ASCENDING)) {
             cq.orderBy(cb.asc(order.get(sortField)));
         }
-        else if (sortOrder.equals(SortOrder.DESCENDING))
-        {
+        else if (sortOrder.equals(SortOrder.DESCENDING)) {
             cq.orderBy(cb.desc(order.get(sortField)));
         }
         List<Predicate> predicateList = new ArrayList<>();
-        for (Map.Entry<String, Object> filter : filters.entrySet())
-        {
-            if (filter.getKey().equals("orderStatus") || filter.getKey().equals(ORDER_ID))
-            {
+        for (Map.Entry<String, Object> filter : filters.entrySet()) {
+            if (filter.getKey().equals("orderStatus") || filter.getKey().equals(ORDER_ID)) {
                 predicateList.add(cb.equal(order.get(filter.getKey()), filter.getValue().toString()));
             }
-            else
-            {
+            else {
                 predicateList.add(cb.like(order.get(filter.getKey()), "%" + filter.getValue().toString() + "%"));
             }
         }
@@ -135,34 +116,33 @@ public class OrdersBean
         cq.where(predicateArray);
 
         TypedQuery<Orders> typedQuery = em.createQuery(select);
-        if (first >= 0)
-        {
+        if (first >= 0) {
             typedQuery.setFirstResult(first);
         }
-        if (pageSize >= 0)
-        {
+        if (pageSize >= 0) {
             typedQuery.setMaxResults(pageSize);
         }
         return typedQuery.getResultList();
     }
-    
-    public Orders addNewOrder(Orders orders)
-    {
+
+    public Orders addNewOrder(Orders orders) {
         em.persist(orders);
         em.flush();
         return orders;
     }
-    
-    public Orders editOrder(Orders orders)
-    {
+
+    public Orders editOrder(Orders orders) {
         orders = em.merge(orders);
         em.flush();
         return orders;
     }
-    
-    public void deleteOrder(Orders orders){
+
+    public Orders deleteOrder(Orders orders) {
         orders = em.find(Orders.class, orders.getOrderId());
-        em.remove(orders);
+        if (orders != null) {
+            em.remove(orders);
+        }
         em.flush();
+        return orders;
     }
 }
